+++ source/MRPch/MRPch.h
@@ -98,7 +98,7 @@
 // in MSVC it dramaticcaly increases the size of PCH file (from 400Mb to 800Mb),
 // but greatly improves compilation time of TUs that include OpenVDB
 #ifdef MR_PCH_USE_OPENVDB
-#include "MRVoxels/MROpenVDB.h"
+#include "MROpenVDB.h"
 #endif
 
 #pragma warning(pop)
