+++ source/MRPch/MRPch.vcxproj
@@ -27,6 +27,7 @@
     <ClInclude Include="MRAsyncLaunchType.h" />
     <ClInclude Include="MREigen.h" />
     <ClInclude Include="MRJson.h" />
+    <ClInclude Include="MROpenVDB.h" />
     <ClInclude Include="MRPch.h" />
     <ClInclude Include="MRTBB.h" />
     <ClInclude Include="MRSpdlog.h" />
