+++ source/MRVoxels/MRVoxels.vcxproj
@@ -73,7 +73,6 @@
     <ClInclude Include="MRMoveMeshToVoxelMaxDeriv.h" />
     <ClInclude Include="MRObjectVoxels.h" />
     <ClInclude Include="MROffset.h" />
-    <ClInclude Include="MROpenVDB.h" />
     <ClInclude Include="MROpenVDBHelper.h" />
     <ClInclude Include="MRPartialOffset.h" />
     <ClInclude Include="MRPointsToDistanceVolume.h" />
